Re-implement taiko's accumulating health processor
