+++ osu.Game/Rulesets/Scoring/HealthProcessor.cs
@@ -2,16 +2,13 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Collections.Generic;
 using osu.Framework.Bindables;
 using osu.Framework.MathUtils;
-using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Judgements;
-using osu.Game.Rulesets.Objects;
 
 namespace osu.Game.Rulesets.Scoring
 {
-    public class HealthProcessor : JudgementProcessor
+    public abstract class HealthProcessor : JudgementProcessor
     {
         /// <summary>
         /// Invoked when the <see cref="ScoreProcessor"/> is in a failed state.
@@ -39,42 +36,14 @@ namespace osu.Game.Rulesets.Scoring
         /// </summary>
         public bool HasFailed { get; private set; }
 
-        private readonly double gameplayStartTime;
+        /// <summary>
+        /// The gameplay start time.
+        /// </summary>
+        protected readonly double GameplayStartTime;
 
-        private IBeatmap beatmap;
-
-        private List<(double time, double health)> healthIncreases;
-        private double targetMinimumHealth;
-        private double drainRate = 1;
-
-        public HealthProcessor(double gameplayStartTime)
+        protected HealthProcessor(double gameplayStartTime)
         {
-            this.gameplayStartTime = gameplayStartTime;
-        }
-
-        public override void ApplyBeatmap(IBeatmap beatmap)
-        {
-            this.beatmap = beatmap;
-
-            healthIncreases = new List<(double time, double health)>();
-            targetMinimumHealth = BeatmapDifficulty.DifficultyRange(beatmap.BeatmapInfo.BaseDifficulty.DrainRate, 0.95, 0.70, 0.30);
-
-            base.ApplyBeatmap(beatmap);
-
-            // Only required during the simulation stage
-            healthIncreases = null;
-        }
-
-        protected override void Update()
-        {
-            base.Update();
-
-            if (!IsBreakTime.Value)
-            {
-                // When jumping from before the gameplay start to after it or vice-versa, we only want to consider any drain since the gameplay start time
-                double lastTime = Math.Max(gameplayStartTime, Time.Current - Time.Elapsed);
-                Health.Value -= drainRate * (Time.Current - lastTime);
-            }
+            GameplayStartTime = gameplayStartTime;
         }
 
         protected override void ApplyResultInternal(JudgementResult result)
@@ -82,13 +51,10 @@ namespace osu.Game.Rulesets.Scoring
             result.HealthAtJudgement = Health.Value;
             result.FailedAtJudgement = HasFailed;
 
-            double healthIncrease = result.Judgement.HealthIncreaseFor(result);
-            healthIncreases?.Add((result.HitObject.GetEndTime() + result.TimeOffset, healthIncrease));
-
             if (HasFailed)
                 return;
 
-            Health.Value += healthIncrease;
+            Health.Value += GetHealthIncreaseFor(result);
 
             if (!DefaultFailCondition && FailConditions?.Invoke(this, result) != true)
                 return;
@@ -104,6 +70,8 @@ namespace osu.Game.Rulesets.Scoring
             // Todo: Revert HasFailed state with proper player support
         }
 
+        protected virtual double GetHealthIncreaseFor(JudgementResult result) => result.Judgement.HealthIncreaseFor(result);
+
         /// <summary>
         /// The default conditions for failing.
         /// </summary>
@@ -112,53 +80,6 @@ namespace osu.Game.Rulesets.Scoring
         protected override void Reset(bool storeResults)
         {
             base.Reset(storeResults);
-
-            drainRate = 1;
-
-            if (storeResults)
-            {
-                int count = 1;
-
-                while (true)
-                {
-                    double currentHealth = 1;
-                    double lowestHealth = 1;
-                    int currentBreak = -1;
-
-                    for (int i = 0; i < healthIncreases.Count; i++)
-                    {
-                        double currentTime = healthIncreases[i].time;
-                        double lastTime = i > 0 ? healthIncreases[i - 1].time : gameplayStartTime;
-
-                        // Subtract any break time from the duration since the last object
-                        if (beatmap.Breaks.Count > 0)
-                        {
-                            while (currentBreak + 1 < beatmap.Breaks.Count && beatmap.Breaks[currentBreak + 1].EndTime < currentTime)
-                                currentBreak++;
-
-                            if (currentBreak >= 0)
-                                lastTime = Math.Max(lastTime, beatmap.Breaks[currentBreak].EndTime);
-                        }
-
-                        // Apply health adjustments
-                        currentHealth -= (healthIncreases[i].time - lastTime) * drainRate;
-                        lowestHealth = Math.Min(lowestHealth, currentHealth);
-                        currentHealth = Math.Min(1, currentHealth + healthIncreases[i].health);
-
-                        // Common scenario for when the drain rate is definitely too harsh
-                        if (lowestHealth < 0)
-                            break;
-                    }
-
-                    if (Math.Abs(lowestHealth - targetMinimumHealth) <= 0.01)
-                        break;
-
-                    count *= 2;
-                    drainRate += 1.0 / count * Math.Sign(lowestHealth - targetMinimumHealth);
-                }
-            }
-
-            healthIncreases.Clear();
 
             Health.Value = 1;
             HasFailed = false;
